source/requirements.txt
# requirements.txt

# API framework
fastapi

uvicorn[standard]

python-multipart

pika

python-dotenv